+++ crates/ruma-federation-api/src/room.rs
@@ -1,0 +1,4 @@
+//! Server room endpoints.
+
+#[cfg(feature = "unstable-msc3843")]
+pub mod report_content;