+++ crates/ruma-federation-api/CHANGELOG.md
@@ -10,6 +10,8 @@ Improvements:
 * Implement `From<SpaceHierarchyParentSummary>` for `SpaceHierarchyChildSummary`
 * Add unstable support for optional `via` field on the `create_invite` endpoint request from
   MSC4125 behind the `unstable-msc4125` feature.
+* Add unstable support for the `report_content` endpoint from MSC3843 behind the
+  `unstable-msc3843` feature.
 
 # 0.8.0